Extract CSV/ZIP logic from reports route into src/lib/reports.ts
Move the CSV content generation and ZIP filename building logic out of
the MonthlyReport component into standalone pure functions:

- buildCsvContent: groups expenses by day/category, produces the full
  CSV string with BOM, Italian headers, and sorted rows.
- buildZipFilename: sanitizes merchant names, resolves extensions from
  content types, and deduplicates filenames via a running count map.

The route component retains all side effects (fetch, toast, file-saver).

Closes #200

Made-with: Cursor
